provider

+++ lib/dto/response/WalletResponse.dart
@@ -1,7 +1,7 @@
 class WalletResponse {
-  final double crypto_balance;
-  final double card_balance;
-  final double balance_amount;
+  final int crypto_balance;
+  final int card_balance;
+  final int balance_amount;
 
   WalletResponse(
       {required this.crypto_balance,
@@ -10,8 +10,8 @@ class WalletResponse {
 
   factory WalletResponse.fromJson(Map<String, dynamic> json) {
     return WalletResponse(
-        card_balance: json['card_balance'] as double,
-        crypto_balance: json['crypto_balance'] as double,
-        balance_amount: json['balance_amount'] as double);
+        card_balance: json['card_balance'] as int,
+        crypto_balance: json['crypto_balance'] as int,
+        balance_amount: json['balance_amount'] as int);
   }
 }
